+++ src/cert/ext.rs
@@ -1,5 +1,6 @@
 //! X509 Extensions
 
+use std::fmt;
 use bcder::{BitString, Captured, Mode, OctetString, Oid, Tag};
 use bcder::{decode, encode};
 use bcder::encode::PrimitiveContent;
@@ -905,6 +906,15 @@ impl UriGeneralName {
 
     pub fn into_rsync_uri(self) -> Option<uri::Rsync> {
         uri::Rsync::from_bytes(self.0.clone()).ok()
+    }
+}
+
+
+//--- Display
+
+impl fmt::Display for UriGeneralName {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", String::from_utf8_lossy(&self.0))
     }
 }